20230816 디자인완료

+++ src/component/MarkList.js
@@ -8,14 +8,22 @@ function MarkList() {
   const [mList, setMList] = useState(data);
   const navigate = useNavigate();
 
+  if(data===null){
+    return
+  }
+
   // 삭제(로컬스토리지에서 삭제하기)
-  const onDeleteHandler = () => {
-    const getID = document.querySelector(".m_id").innerHTML;
-    console.log(getID);
-    let reData = data.filter((m) => m !== getID);
-    console.log("MarkList reData", reData);
-    localStorage.setItem("Pg_CM", JSON.stringify(reData));
-    setMList(reData);
+  const onDeleteHandler = (e) => {
+    e.stopPropagation()
+    if(window.confirm('삭제하시겠습니까?')){
+      const getID = document.querySelector(".m_id").innerHTML;
+      console.log(getID);
+      let reData = data.filter((m) => m !== getID);
+      // 삭제후 마지막이 되면 빈칸저장시킴
+      console.log("MarkList reData", reData);
+      localStorage.setItem("Pg_CM", JSON.stringify(reData));
+      setMList(reData);
+    }
   };
 
   // 박물관 행 선택시 디테일화면으로 이동
@@ -24,7 +32,7 @@ function MarkList() {
     navigate("/museum/" + getID);
   };
 
-  return data.map((mark) => {
+  return mList.map((mark) => {
     // museum에서 찾기
     const _museum = museum;
     const Mdata = _museum.find((m) => String(m.id) === String(mark));
@@ -37,7 +45,8 @@ function MarkList() {
         <td data-th="지역">{Mdata.area_adr}</td>
         <td data-th="박물관">{Mdata.mus_nam}</td>
         <td className="btn_del">
-          <button onClick={onDeleteHandler}>삭제</button>
+          <button className="mark-btn-del" onClick={onDeleteHandler}>
+            <i class="fa-solid fa-xmark"/></button>
         </td>
       </tr>
     );
